Fix Mobile UI for task detail
