--- src/components/task-detail/task-body-left/style.module.scss
@@ -1,28 +1,0 @@
-.task-body-left :global {
-  @apply basis-2/3;
-  .inner {
-    @apply space-y-5 divide-y;
-  }
-  .title {
-    @apply mt-5 mb-3 flex items-center gap-2.5 first:mt-0;
-    i::before {
-      @apply text-[20px];
-    }
-    h4 {
-      @apply text-h4 font-medium text-slate-700;
-    }
-  }
-  .edit-btn {
-    @apply rounded bg-slate-200 p-1 text-h6 text-slate-500;
-  }
-
-  textarea {
-    box-shadow: none;
-  }
-  .task-images {
-    @apply flex flex-col gap-3;
-    &-upload {
-      @apply mt-3 flex gap-3 overflow-x-auto bg-blue-100 p-2;
-    }
-  }
-}
