--- src/components/task-detail/task-body-right/due-date/index.tsx
@@ -1,11 +1,0 @@
-import {FC} from 'react';
-
-const DueDate: FC = () => {
-  return (
-    <div className="due-date">
-      <p className="title">Due date</p>
-      <p>4/11/2022</p>
-    </div>
-  );
-};
-export default DueDate;
