--- src/components/task-detail/task-body-right/priority/style.module.scss
@@ -1,10 +1,0 @@
-.priority :global {
-  .MuiInputBase-root {
-    .MuiSelect-select {
-      @apply p-0;
-    }
-    fieldset {
-      @apply hidden;
-    }
-  }
-}
