--- src/components/task-detail/task-body-right/start-date/index.tsx
@@ -1,11 +1,0 @@
-import {FC} from 'react';
-
-const StartDate: FC = () => {
-  return (
-    <div className="start-date">
-      <p className="title">Start date</p>
-      <p>3/11/2022</p>
-    </div>
-  );
-};
-export default StartDate;
