--- src/components/task-detail/task-body-right/style.module.scss
@@ -1,20 +1,0 @@
-.task-body-right :global {
-  @apply basis-1/3;
-  .container {
-    @apply space-y-5 divide-y;
-  }
-
-  .title {
-    @apply py-3 text-h5 font-semibold;
-  }
-
-  .assignee {
-    .user {
-      @apply flex gap-x-3;
-    }
-  }
-
-  .time-state {
-    @apply space-y-2 py-2;
-  }
-}
