+++ src/components/task-detail/task-body/body-left/attachment/images/style.module.scss
@@ -1,0 +1,29 @@
+.task-images :global {
+  .task-image {
+    @apply flex gap-2.5 rounded-lg hover:bg-white;
+
+    .image {
+      @apply relative h-20 w-28 overflow-hidden rounded border bg-white;
+    }
+    .info {
+      @apply h-full;
+      &-name {
+        @apply flex break-all text-slate-700;
+        span {
+          @apply text-h6 font-medium;
+        }
+      }
+      &-date {
+        @apply my-1 text-slate-600;
+      }
+      &-actions {
+        @apply flex gap-2.5 text-h7 text-slate-500 underline;
+      }
+    }
+    &.upload {
+      .image {
+        @apply h-16 w-24;
+      }
+    }
+  }
+}
