+++ src/components/task-detail/task-body/body-right/date/start-date/index.tsx
@@ -1,0 +1,12 @@
+import classNames from 'classnames';
+import {FC, HTMLAttributes} from 'react';
+
+const StartDate: FC<HTMLAttributes<HTMLDivElement>> = ({className}) => {
+  return (
+    <div className={classNames('start-date', className)}>
+      <p className="title">Start date</p>
+      <p>3/11/2022</p>
+    </div>
+  );
+};
+export default StartDate;
