--- src/components/task-detail/task-description/style.module.scss
@@ -1,3 +1,0 @@
-.task-description :global {
-  @apply block;
-}
